+++ src/thread/sync/PosixMutex.h
@@ -19,12 +19,12 @@ namespace pipey {
 
 				
 #ifdef _POSIX_THREAD_PROCESS_SHARED
-				POSIX_MUTEX_INIT(bool processShared, int type) :
+				POSIX_MUTEX_INIT(bool processShared, int32_t type) :
 				bProcessShared(processShared),
 				nType(type)
 				{ };
 #else
-				POSIX_MUTEX_INIT(int type) :
+				POSIX_MUTEX_INIT(int32_t type) :
 				nType(type)
 				{ };
 #endif
@@ -32,7 +32,7 @@ namespace pipey {
 #ifdef _POSIX_THREAD_PROCESS_SHARED
 				bool bProcessShared;
 #endif
-				int nType;
+				int32_t nType;
 			};
 
 			class CPosixMutex : public ITriableTimerableLock {
@@ -58,7 +58,7 @@ namespace pipey {
 
 				virtual pipey::common::SYNC_RESULT TryLock();
 
-				virtual pipey::common::SYNC_RESULT AcquireTimedLock(unsigned long nMilliSeconds = pipey::common::TIME_INFINITE);
+				virtual pipey::common::SYNC_RESULT AcquireTimedLock(uint32_t nMilliSeconds = pipey::common::TIME_INFINITE);
 
 				virtual pthread_mutex_t *GetHandle();
