+++ src/util/PosixTimeHelper.cpp
@@ -3,23 +3,23 @@
 namespace pipey {
 	namespace util {
 		
-		unsigned long timeval_ext::to_milli_seconds()
+		uint64_t timeval_ext::to_milli_seconds()
 		{
-			return ((unsigned long)tv_sec)*1000 + ((unsigned long)tv_usec)/1000;
+			return tv_sec*1000 + tv_usec/1000;
 		}
 
-		void RelativeToAbsolute(unsigned long nMilliseconds, timespec &rAbsolute)
+		void RelativeToAbsolute(uint32_t nMilliseconds, timespec &rAbsolute)
 		{
 			timeval now;
 			gettimeofday(&now, NULL);
 
-			long milli = (nMilliseconds%1000) + (now.tv_usec/1000);
+			uint64_t milli = (nMilliseconds%1000) + (now.tv_usec/1000);
 
 			rAbsolute.tv_sec = milli/1000 + now.tv_sec + nMilliseconds/1000;
 			rAbsolute.tv_nsec = (milli%1000) * 1000 * 1000;
 		}
 
-		unsigned long GetTimeMilliSeconds()
+		uint64_t GetTimeMilliSeconds()
 		{
 			timeval_ext now;
 			gettimeofday(&now, NULL);
